Remove dtype from confusion_matrix initialization.
Changed the confusion matrix initialization to avoid specifying the data type explicitly. This ensures the matrix can handle float values if necessary, providing more flexibility in subsequent analysis.

+++ supervised_learning/error_analysis/0-create_confusion.py
@@ -17,7 +17,7 @@ def create_confusion_matrix(labels, logits):
 
     classes = labels.shape[1]
 
-    confusion_matrix = np.zeros((classes, classes), dtype=int)
+    confusion_matrix = np.zeros((classes, classes))
 
     actual = np.argmax(labels, axis=1)
     predictions = np.argmax(logits, axis=1)
